Correções em category e comments para esses funcionarem

+++ blog/views.py
@@ -1,24 +1,38 @@
-from django.shortcuts import render, get_object_or_404, redirect 
+from django.shortcuts import render, get_object_or_404, redirect
 from .models import Post, Comment, Category
 from .forms import PostForm
-from django.urls import reverse
+from django.contrib.auth.decorators import login_required
 
 def post_list(request):
     posts = Post.objects.all() 
     return render(request, 'blog/post_list.html', {'posts': posts})
+
 def post_detail(request, pk):
-    post = get_object_or_404(Post, pk=pk) 
-    return render(request, 'blog/post_detail.html', {'post': post})
-def post_new(request):
+    post = get_object_or_404(Post, pk=pk)
+    comments = post.comments.all()
     if request.method == 'POST':
+        form = CommentForm(request.POST)
+        if form.is_valid():
+            comment = form.save(commit=False)
+            comment.post = post
+            comment.author = request.user
+            comment.save()
+            return redirect('post_detail', pk=post.pk)
+    else:
+        form = CommentForm()
+    return render(request, 'blog/post_detail.html', {'post': post, 'comments': comments, 'form': form})
+def post_new(request):
+    if request.method == "POST":
         form = PostForm(request.POST)
         if form.is_valid():
-            form.save()
-            return redirect('post_list')
+            post = form.save(commit=False)
+            post.save()
+            form.save_m2m()
+            return redirect('post_detail', pk=post.pk)
     else:
         form = PostForm()
-    return render(request, 'blog/post_form.html', {'form': form})
-    
+    return render(request, 'blog/post_form2.html', {'form': form})
+
 def post_edit(request, pk):
     post = get_object_or_404(Post, pk=pk)
     if request.method == 'POST':
@@ -29,24 +43,32 @@ def post_edit(request, pk):
     else:
         form = PostForm(instance=post)
     return render(request, 'blog/post_form.html', {'form': form})
-from django.shortcuts import render, redirect, get_object_or_404
-from .models import Post, Comment
+
+from .forms import CommentForm
 
 def add_comment(request, post_pk):
     post = get_object_or_404(Post, pk=post_pk)
 
     if request.method == 'POST':
-        text = request.POST.get('text')
-        if text:
-            Comment.objects.create(post=post, author=None, text=text)
+        form = CommentForm(request.POST)
+        if form.is_valid():
+            comment = form.save(commit=False)
+            comment.post = post
+            comment.author = request.user
+            comment.save()
             return redirect('post_detail', pk=post.pk)
+    else:
+        form = CommentForm()
 
-    return render(request, 'blog/post_detail.html', {'post': post})
-def post_delete(request, pk):
+    comments = post.comments.all()
+    return render(request, 'blog/post_detail.html', {'post': post, 'form': form, 'comments': comments})
+
+def post_delete_confirm(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.delete()
-    return redirect(reverse('post_list'))
-
+    if request.method == 'POST':
+        post.delete()
+        return redirect('post_list')
+    return render(request, 'blog/post_delete_confirm.html', {'post': post})
 
 def category_list(request):
     categories = Category.objects.all()
@@ -54,4 +76,5 @@ def category_list(request):
 
 def category_detail(request, category_id):
     category = get_object_or_404(Category, id=category_id)  
+    posts = category.posts.all()
     return render(request, 'blog/category_detail.html', {'category': category, 'posts': posts})
